fix(frame): str split_get/rsplit_get/split_regex_get support negative index
pandas Series.str.split(pat).str[n] (and rsplit/regex variants) accept a negative
n that counts from the end, e.g. split('-').str[-1] is the last part. FP's three
split-get methods took n: usize, so negatives were impossible.

Change them to n: i64 and add split_part_at, which resolves a negative n as
len + n and returns a missing value for any out-of-range index (matching pandas).
The three fp-conformance call sites pass their usize fixture arg as n as i64.
Verified vs live pandas 2.2.3 ('a-b-c': [-1]='c', [-2]='b', [-3]='a', [-4]=NaN);
fp-frame str 16/0, clippy clean, fp-conformance compiles. Closes
br-frankenpandas-7evu8.

+++ crates/fp-conformance/src/lib.rs
@@ -16728,7 +16728,7 @@ fn execute_series_module_utility_fixture_operation(
                 .ok_or_else(|| "str_split_n required for series_str_split_get".to_owned())?;
             series
                 .str()
-                .split_get(pat, n)
+                .split_get(pat, n as i64)
                 .map_err(|err| err.to_string())
         }
         FixtureOperation::SeriesStrJoin => {
@@ -16787,7 +16787,7 @@ fn execute_series_module_utility_fixture_operation(
             let n = fixture.str_split_n.unwrap_or(0);
             series
                 .str()
-                .rsplit_get(pat, n)
+                .rsplit_get(pat, n as i64)
                 .map_err(|err| err.to_string())
         }
         FixtureOperation::SeriesStrSplitRegexGet => {
@@ -16795,7 +16795,7 @@ fn execute_series_module_utility_fixture_operation(
             let n = fixture.str_split_n.unwrap_or(0);
             series
                 .str()
-                .split_regex_get(pat, n)
+                .split_regex_get(pat, n as i64)
                 .map_err(|err| err.to_string())
         }
         FixtureOperation::SeriesStrCountMatches => {
